Server/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
azure-core
